+++ lib/features/collections/widgets/collection_table/table_column.dart
@@ -6,6 +6,7 @@ enum TableColumn {
   status,
   tag,
   rating,
+  externalRating,
   year,
   added,
 }